src/requirements.txt
graphene_mongo
flask
flask_graphql
graphene
gunicorn
pandas
flask_mongoengine